+++ xray/core/dataset.py
@@ -1247,6 +1247,61 @@ class Dataset(Mapping, common.ImplementsDatasetReduce):
         """
         return common.squeeze(self, self.dims, dim)
 
+    def dropna(self, dim, how='any', thresh=None, subset=None):
+        """Returns a new dataset with dropped labels for missing values along
+        the provided dimension.
+
+        Parameters
+        ----------
+        dim : str
+            Dimension along which to drop missing values. Dropping along
+            multiple dimensions simultaneously is not yet supported.
+        how : {'any', 'all'}, optional
+            * any : if any NA values are present, drop that label
+            * all : if all values are NA, drop that label
+        thresh : int, default None
+            If supplied, require this many non-NA values.
+        subset : sequence, optional
+            Subset of variables to check for missing values. By default, all
+            variables in the dataset are checked.
+
+        Returns
+        -------
+        Dataset
+        """
+        # TODO: consider supporting multiple dimensions? Or not, given that
+        # there are some ugly edge cases, e.g., pandas's dropna differs
+        # depending on the order of the supplied axes.
+
+        if dim not in self.dims:
+            raise ValueError('%s must be a single dataset dimension' % dim)
+
+        if subset is None:
+            subset = list(self.vars)
+
+        count = np.zeros(self.dims[dim], dtype=int)
+        size = 0
+
+        for k in subset:
+            array = self._arrays[k]
+            if dim in array.dims:
+                dims = [d for d in array.dims if d != dim]
+                count += array.count(dims)
+                size += np.prod([self.dims[d] for d in dims])
+
+        if thresh is not None:
+            mask = count >= thresh
+        elif how == 'any':
+            mask = count == size
+        elif how == 'all':
+            mask = count > 0
+        elif how is not None:
+            raise ValueError('invalid how option: %s' % how)
+        else:
+            raise TypeError('must specify how or thresh')
+
+        return self.isel(**{dim: mask})
+
     def reduce(self, func, dim=None, keep_attrs=False, **kwargs):
         """Reduce this dataset by applying `func` along some dimension(s).